files/anatomy/bone/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
httpx>=0.27
PyYAML>=6.0
# Track B (2026-04-26): JWT validation against Authentik JWKS for
# operational endpoints. PyJWT[crypto] pulls in `cryptography` for RS256.
PyJWT[crypto]>=2.9
